+++ projects/ui-kit/forms/ui-swatch-picker/src/public-api.ts
@@ -1,0 +1,5 @@
+/*
+ * Public API Surface of ui-kit/ui-swatch-picker
+ */
+
+export * from './lib/ui-swatch-picker';